Refactor utils into a subpackage
